Windows 10 joins the party
At least at first, we are going to be supporting Win32 for Steam and UWP
for Windows App Store.

Steam provides a match making and lobbying service, among other things
like Account Management, all for free via their SteamWorks SDK. In the
App world (UWP, Android, and iOS), we have no such conveniences, so we
would need to use a service like Google Play Game Services to get these
platforms connected. The only platforms explicitly supported via an SDK
is Android. The iOS one is broken currently. So for UWP, we would need
to write all of our own UI to interface with Google's REST service. So
needless to say, the app versions of the game are going to be lower
priority. Project Dante is going to be a PC > Console > Mobile focused
game.